--- test/09_setup_cfg/setup.cfg
@@ -1,6 +1,0 @@
-[metadata]
-name = spam
-version = attr: spam.__version__
-
-[options]
-packages = find:
