--- test/09_setup_cfg/setup.py
@@ -1,8 +1,0 @@
-import os
-
-from setuptools import setup, Extension
-
-
-setup(
-    ext_modules=[Extension('spam.spam', sources=['spam/spam.c'])],
-)
