+++ test/cpp_project_template/spam.cpp
@@ -1,34 +1,17 @@
 #include <Python.h>
-#include <malloc.h>
+#include <string>
 
-#if !defined(__GLIBC_PREREQ)
-#error "Must run on a glibc linux environment"
-#endif
-
-#if !__GLIBC_PREREQ(2, 5)  /* manylinux1 is glibc 2.5 */
-#error "Must run on a glibc >= 2.5 linux environment"
-#endif
+{{spam_cpp_top_level_add}}
 
 static PyObject *
 spam_system(PyObject *self, PyObject *args)
 {
     const char *command;
-    int sts = 0;
+    int sts;
 
     if (!PyArg_ParseTuple(args, "s", &command))
         return NULL;
-
-#if defined(__GLIBC_PREREQ) && __GLIBC_PREREQ(2, 17)  /* manylinux2014 is glibc 2.17 */
-    // secure_getenv is only available in manylinux2014, ensuring
-    // that only a manylinux2014 wheel is produced
-    sts = (int)secure_getenv("NON_EXISTING_ENV_VARIABLE");
-#elif defined(__GLIBC_PREREQ) && __GLIBC_PREREQ(2, 10)  /* manylinux2010 is glibc 2.12 */
-    // malloc_info is only available on manylinux2010+
-    sts = malloc_info(0, stdout);
-#endif
-    if (sts == 0) {
-        sts = system(command);
-    }
+    sts = system(command);
     return PyLong_FromLong(sts);
 }
 
@@ -49,7 +32,7 @@ spam_system(PyObject *self, PyObject *args)
 #endif
 
 static PyMethodDef module_methods[] = {
-    {"system", (PyCFunction)spam_system, METH_VARARGS,
+    {"system", (PyCFunction)spam_system, METH_VARARGS, 
      "Execute a shell command."},
     {NULL}  /* Sentinel */
 };
@@ -58,9 +41,9 @@ MOD_INIT(spam)
 {
     PyObject* m;
 
-    MOD_DEF(m,
-            "spam",
-            "Example module",
+    MOD_DEF(m, 
+            "spam", 
+            "Example module", 
             module_methods,
             -1)
