+++ test/test_05_environment.py
@@ -30,7 +30,7 @@ def test(tmpdir):
     # write some information into the CIBW_ENVIRONMENT, for expansion and
     # insertion into the environment by cibuildwheel. This is checked
     # in setup.py
-    utils.cibuildwheel_run(
+    actual_wheels = utils.cibuildwheel_run(
         project_dir,
         add_env={
             "CIBW_ENVIRONMENT": """CIBW_TEST_VAR="a b c" CIBW_TEST_VAR_2=1 CIBW_TEST_VAR_3="$(echo 'test string 3')" PATH=$PATH:/opt/cibw_test_path""",
@@ -40,5 +40,4 @@ def test(tmpdir):
 
     # also check that we got the right wheels built
     expected_wheels = utils.expected_wheels("spam", "0.1.0")
-    actual_wheels = os.listdir("wheelhouse")
     assert set(actual_wheels) == set(expected_wheels)
